backend/requirements.txt
amqp==5.0.6
asgiref==3.3.4
astroid==2.5.7
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dateutils==0.6.12
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
isort==5.8.0
kombu==5.1.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
nsetools==1.0.11
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
